storage_backend_s3: fix other region name handling
