+++ storage_backend_s3/tests/__init__.py
@@ -1,2 +1,1 @@
 from . import test_amazon_s3
-from . import test_fake_s3
